Murali Pattathe
This is the patch for the OleConvertOLESTREAMToIStorage wine API.
This API calls another function called OLECONVERT_LoadOLE10 which
basically read the OLE stream from the file through the callback
function (this is passed by the application).

The reading format was wrong. Still I am not very sure this is 100 %
correct format. I did some reverse engineering and found the format. I
have tried with different OLE object and its worked in paradox. I
haven't tested with any other apps.