kernel32: Fix the GetComputerName* functions to not write to the buffer at all if there is not enough space.
Add tests for the GetComputerName* functions that show this issue.

Fix an off-by-one error in setting the size when there is not enough space.

Remove the exception handlers for GetComputerNameW,
GetComputerNameExA/W, since testing on XP SP2 shows that access
violations are not caught and these are not present on Win9x, which
does have an exception handler.