notepad: Improve printing considerably.
Make fonts and the other measurements the correct size in relation to
the actual printer resolution instead of thinking it uses the same
resolution as the display.  Before fixed sized (and for most printer
resolutions way to small) margins where used and the font size was
taken directly from the screen font independent of the actual printer
resolution, resulting in a completely unreadable micro text on most
printouts.