mshtml: Fix inserting a hyperlink when text is already selected.
There is no need to use cmd_insertLinkNoUI when we have the code to do 
most of the action anyway. Make use of the handy 
InsertLinkAroundSelection to insert the "a" element when not inserting 
at the caret and don't create a text node in this case because it is 
redundant.