user32: Fix MonitorFromRect to cope with the absence of the MONITOR_DEFAULTTONEAREST flag.
Previously, the code would return any monitor found, regardless of 
whether it intersected the given rect or was the nearest monitor. This 
is fixed by adding a new flag that causes monitor_enum to only find the 
nearest monitor if MONITOR_DEFAULTTONEAREST is specified.

Also add a trace for MonitorFromWindow, since it is called in many 
places within user32 and so can't be traced using a relay trace.