comctl32: rebar: Don't assume the first band is visible.
The code required that a row started with a visible band. However this
was not met if the first band of the rebar was hidden. The new code
should fix this case. I've also made it explicit in the comments that
the functions require the first band to be visible and renamed
next_band, prev_band to next_visible and prev_visible to make it clear
what these functions do.