wordpad: Added style to the richedit control to not hide selection.
When run with the native richedit control the selection will be hidden
when focus is lost without the ES_NOHIDESEL style applied to the
richedit control.  This can be annoying in native windows when focus is
on a combobox in the toolbar, because it won't show what text will be
changed.  This patch adds the ES_NOHIDESEL style to be consistent with
native wordpad.