wordpad: Apply changes on dropdown selection for comboboxes on toolbar.
The changes were previously being applied only when the combobox's edit
control lost focus notifying an end edit event.  In addition, the
selection of an item from the dropdown list should also cause the
changes to be applied. New functions were created to handle the
duplication in expected behaviour for these two events.