winemine: Set mines after first choice.
The first choice never is on a mine in Windows.  This can be tested by
making a custom game with as many mines as possible, and then playing
the start of many games, and the first choice will never be on a mine.

This is done to make the game reasonable, since after the first choice
there will at least be some information given to make the next choice.