winemine: The number of mines should be visible before first choice.
There were a few lines of code that I removed that explicitly kept the
number of mines and the timer unlit before the game started.  This isn't
consistent with Windows, so I don't know what the purpose was for doing
this.  In fact, it prevents the number of mines from being shown ahead
of time, which is useful information.