user32: Increased area for scrolling via mouse drag outside scrollbar.
When dragging the scrollbar thumb with the mouse, the mouse is able to
move away from the scrollbar and keep scrolling so long as it isn't too
far away from the scrollbar.  This makes it easier to quickly scroll
with the mouse.

All that this patch changes is the distance that the mouse can be moved
away from the scrollbar before it is consider outside of the scrollbar
and returns to its original position.  The distances are proportional to
the size of the scrollbar.