user32/tests: Don't check menu items when testing MNS_CHECKORBMP.
If the MNS_CHECKORBMP style is set on a menu, and the menu is checked,
the bitmap is not drawn. For the menu tests, that means we get no
WM_DRAWITEM message for an owner-drawn bitmap, and the bounding box of
the bitmap is not tested. We should leave the menu item unchecked in
this case so we can test it.